+++ src/publishers.cpp
@@ -372,6 +372,9 @@ const char* ToString(Dataset dataset) {
     case Dataset::DbeqMax: {
       return "DBEQ.MAX";
     }
+    case Dataset::XnasBasic: {
+      return "XNAS.BASIC";
+    }
     default: {
       return "Unknown";
     }
@@ -474,6 +477,9 @@ Dataset FromString(const std::string& str) {
   }
   if (str == "DBEQ.MAX") {
     return Dataset::DbeqMax;
+  }
+  if (str == "XNAS.BASIC") {
+    return Dataset::XnasBasic;
   }
   throw InvalidArgumentError{"FromString<Dataset>", "str",
                              "unknown value '" + str + '\''};
@@ -720,6 +726,15 @@ Venue PublisherVenue(Publisher publisher) {
     }
     case Publisher::DbeqMaxLtse: {
       return Venue::Ltse;
+    }
+    case Publisher::XnasBasicXnas: {
+      return Venue::Xnas;
+    }
+    case Publisher::XnasBasicFinn: {
+      return Venue::Finn;
+    }
+    case Publisher::XnasBasicFinc: {
+      return Venue::Finc;
     }
     default: {
       throw InvalidArgumentError{
@@ -971,6 +986,15 @@ Dataset PublisherDataset(Publisher publisher) {
     case Publisher::DbeqMaxLtse: {
       return Dataset::DbeqMax;
     }
+    case Publisher::XnasBasicXnas: {
+      return Dataset::XnasBasic;
+    }
+    case Publisher::XnasBasicFinn: {
+      return Dataset::XnasBasic;
+    }
+    case Publisher::XnasBasicFinc: {
+      return Dataset::XnasBasic;
+    }
     default: {
       throw InvalidArgumentError{
           "PublisherDataset", "publisher",
@@ -1221,6 +1245,15 @@ const char* ToString(Publisher publisher) {
     }
     case Publisher::DbeqMaxLtse: {
       return "DBEQ.MAX.LTSE";
+    }
+    case Publisher::XnasBasicXnas: {
+      return "XNAS.BASIC.XNAS";
+    }
+    case Publisher::XnasBasicFinn: {
+      return "XNAS.BASIC.FINN";
+    }
+    case Publisher::XnasBasicFinc: {
+      return "XNAS.BASIC.FINC";
     }
     default: {
       return "Unknown";
@@ -1474,6 +1507,15 @@ Publisher FromString(const std::string& str) {
   }
   if (str == "DBEQ.MAX.LTSE") {
     return Publisher::DbeqMaxLtse;
+  }
+  if (str == "XNAS.BASIC.XNAS") {
+    return Publisher::XnasBasicXnas;
+  }
+  if (str == "XNAS.BASIC.FINN") {
+    return Publisher::XnasBasicFinn;
+  }
+  if (str == "XNAS.BASIC.FINC") {
+    return Publisher::XnasBasicFinc;
   }
   throw InvalidArgumentError{"FromString<Publisher>", "str",
                              "unknown value '" + str + '\''};